refactor: browser type
